+++ modules/aks/main.tf
@@ -12,6 +12,7 @@ resource "azurerm_kubernetes_cluster" "this" {
   workload_identity_enabled = var.workload_identity_enabled
   automatic_upgrade_channel = var.automatic_upgrade_channel != "none" ? var.automatic_upgrade_channel : null
   node_resource_group       = var.node_resource_group_name
+  private_dns_zone_id       = var.private_dns_zone_id
 
   default_node_pool {
     name                        = "system"
@@ -35,6 +36,29 @@ resource "azurerm_kubernetes_cluster" "this" {
     }
   }
 
+  dynamic "auto_scaler_profile" {
+    for_each = var.auto_scaler_profile != null ? [var.auto_scaler_profile] : []
+    content {
+      balance_similar_node_groups      = auto_scaler_profile.value.balance_similar_node_groups
+      empty_bulk_delete_max            = auto_scaler_profile.value.empty_bulk_delete_max
+      expander                         = auto_scaler_profile.value.expander
+      max_graceful_termination_sec     = auto_scaler_profile.value.max_graceful_termination_sec
+      max_node_provisioning_time       = auto_scaler_profile.value.max_node_provisioning_time
+      max_unready_nodes                = auto_scaler_profile.value.max_unready_nodes
+      max_unready_percentage           = auto_scaler_profile.value.max_unready_percentage
+      new_pod_scale_up_delay           = auto_scaler_profile.value.new_pod_scale_up_delay
+      scale_down_delay_after_add       = auto_scaler_profile.value.scale_down_delay_after_add
+      scale_down_delay_after_delete    = auto_scaler_profile.value.scale_down_delay_after_delete
+      scale_down_delay_after_failure   = auto_scaler_profile.value.scale_down_delay_after_failure
+      scale_down_unneeded              = auto_scaler_profile.value.scale_down_unneeded
+      scale_down_unready               = auto_scaler_profile.value.scale_down_unready
+      scale_down_utilization_threshold = auto_scaler_profile.value.scale_down_utilization_threshold
+      scan_interval                    = auto_scaler_profile.value.scan_interval
+      skip_nodes_with_local_storage    = auto_scaler_profile.value.skip_nodes_with_local_storage
+      skip_nodes_with_system_pods      = auto_scaler_profile.value.skip_nodes_with_system_pods
+    }
+  }
+
   identity {
     type = "SystemAssigned"
   }
@@ -47,6 +71,59 @@ resource "azurerm_kubernetes_cluster" "this" {
     service_cidr        = var.network_profile.service_cidr
     dns_service_ip      = var.network_profile.dns_service_ip
     outbound_type       = var.network_profile.outbound_type
+
+    dynamic "load_balancer_profile" {
+      for_each = var.network_profile.load_balancer_profile != null ? [var.network_profile.load_balancer_profile] : []
+      content {
+        managed_outbound_ip_count = load_balancer_profile.value.managed_outbound_ip_count
+        outbound_ip_address_ids   = load_balancer_profile.value.outbound_ip_address_ids
+        outbound_ip_prefix_ids    = load_balancer_profile.value.outbound_ip_prefix_ids
+        outbound_ports_allocated  = load_balancer_profile.value.outbound_ports_allocated
+        idle_timeout_in_minutes   = load_balancer_profile.value.idle_timeout_in_minutes
+      }
+    }
+  }
+
+  dynamic "maintenance_window" {
+    for_each = var.maintenance_window != null ? [var.maintenance_window] : []
+    content {
+      dynamic "allowed" {
+        for_each = maintenance_window.value.allowed
+        content {
+          day   = allowed.value.day
+          hours = allowed.value.hours
+        }
+      }
+      dynamic "not_allowed" {
+        for_each = maintenance_window.value.not_allowed != null ? maintenance_window.value.not_allowed : []
+        content {
+          start = not_allowed.value.start
+          end   = not_allowed.value.end
+        }
+      }
+    }
+  }
+
+  dynamic "maintenance_window_auto_upgrade" {
+    for_each = var.maintenance_window_auto_upgrade != null ? [var.maintenance_window_auto_upgrade] : []
+    content {
+      frequency    = maintenance_window_auto_upgrade.value.frequency
+      interval     = maintenance_window_auto_upgrade.value.interval
+      duration     = maintenance_window_auto_upgrade.value.duration
+      day_of_week  = maintenance_window_auto_upgrade.value.day_of_week
+      day_of_month = maintenance_window_auto_upgrade.value.day_of_month
+      week_index   = maintenance_window_auto_upgrade.value.week_index
+      start_time   = maintenance_window_auto_upgrade.value.start_time
+      utc_offset   = maintenance_window_auto_upgrade.value.utc_offset
+      start_date   = maintenance_window_auto_upgrade.value.start_date
+      dynamic "not_allowed" {
+        for_each = maintenance_window_auto_upgrade.value.not_allowed != null ? maintenance_window_auto_upgrade.value.not_allowed : []
+        content {
+          start = not_allowed.value.start
+          end   = not_allowed.value.end
+        }
+      }
+    }
   }
 
   azure_active_directory_role_based_access_control {